+++ src/main/kotlin/com/sexysisters/tojserverv2/domain/school/SchoolReader.kt
@@ -1,0 +1,5 @@
+package com.sexysisters.tojserverv2.domain.school
+
+interface SchoolReader {
+    fun getSchool(code: String): School
+}
