+++ src/main/kotlin/com/sexysisters/tojserverv2/domain/school/SchoolStore.kt
@@ -1,0 +1,5 @@
+package com.sexysisters.tojserverv2.domain.school
+
+interface SchoolStore {
+    fun store(school: School): School
+}
