--- src/main/kotlin/com/sexysisters/tojserverv2/domain/school/design/SchoolReader.kt
@@ -1,7 +1,0 @@
-package com.sexysisters.tojserverv2.domain.school.design
-
-import com.sexysisters.tojserverv2.domain.school.School
-
-interface SchoolReader {
-    fun getSchool(code: String): School
-}
