--- src/main/kotlin/com/sexysisters/tojserverv2/domain/school/design/SchoolStore.kt
@@ -1,7 +1,0 @@
-package com.sexysisters.tojserverv2.domain.school.design
-
-import com.sexysisters.tojserverv2.domain.school.School
-
-interface SchoolStore {
-    fun store(school: School): School
-}
