+++ src/main/kotlin/com/sexysisters/tojserverv2/domain/school/service/SchoolServiceImpl.kt
@@ -3,13 +3,12 @@ package com.sexysisters.tojserverv2.domain.school.service
 import com.sexysisters.tojserverv2.domain.school.SchoolCommand
 import com.sexysisters.tojserverv2.domain.school.SchoolInfo
 import com.sexysisters.tojserverv2.domain.school.SchoolMapper
-import com.sexysisters.tojserverv2.domain.school.design.SchoolReader
-import com.sexysisters.tojserverv2.domain.school.design.SchoolStore
-import com.sexysisters.tojserverv2.domain.school.design.StudentReader
+import com.sexysisters.tojserverv2.domain.school.SchoolReader
+import com.sexysisters.tojserverv2.domain.school.SchoolStore
+import com.sexysisters.tojserverv2.domain.student.StudentReader
 import com.sexysisters.tojserverv2.domain.school.exception.SchoolException
-import com.sexysisters.tojserverv2.domain.school.student.Status
-import com.sexysisters.tojserverv2.domain.school.student.Student
-import com.sexysisters.tojserverv2.domain.school.student.updateStatus
+import com.sexysisters.tojserverv2.domain.student.Status
+import com.sexysisters.tojserverv2.domain.student.updateStatus
 import com.sexysisters.tojserverv2.infrastructure.neis.NeisSchoolReader
 import org.springframework.stereotype.Service
 import org.springframework.transaction.annotation.Transactional
