+++ src/main/kotlin/com/sexysisters/tojserverv2/domain/student/StudentReader.kt
@@ -1,0 +1,6 @@
+package com.sexysisters.tojserverv2.domain.student
+
+interface StudentReader {
+    fun getStudent(id: Long): Student
+    fun getCurrentStudent(): Student
+}
