+++ src/main/kotlin/com/sexysisters/tojserverv2/domain/student/StudentStore.kt
@@ -1,0 +1,5 @@
+package com.sexysisters.tojserverv2.domain.student
+
+interface StudentStore {
+    fun store(student: Student): Long
+}
