+++ src/main/kotlin/com/sexysisters/tojserverv2/domain/student/service/StudentService.kt
@@ -1,4 +1,4 @@
-package com.sexysisters.tojserverv2.domain.school.service
+package com.sexysisters.tojserverv2.domain.student.service
 
 import com.sexysisters.tojserverv2.domain.school.SchoolCommand
