+++ src/main/kotlin/com/sexysisters/tojserverv2/infrastructure/school/SchoolReaderImpl.kt
@@ -1,7 +1,7 @@
 package com.sexysisters.tojserverv2.infrastructure.school
 
 import com.sexysisters.tojserverv2.domain.school.School
-import com.sexysisters.tojserverv2.domain.school.design.SchoolReader
+import com.sexysisters.tojserverv2.domain.school.SchoolReader
 import com.sexysisters.tojserverv2.domain.school.exception.SchoolException
 import org.springframework.stereotype.Component
 import org.springframework.transaction.annotation.Transactional
