+++ src/main/kotlin/com/sexysisters/tojserverv2/infrastructure/school/student/StudentRepository.kt
@@ -1,6 +1,6 @@
 package com.sexysisters.tojserverv2.infrastructure.school.student
 
-import com.sexysisters.tojserverv2.domain.school.student.Student
+import com.sexysisters.tojserverv2.domain.student.Student
 import org.springframework.data.jpa.repository.JpaRepository
 
 interface StudentRepository : JpaRepository<Student, Long>
